move skill tree to components folder
