+++ apps/web/src/pages/Applications/ApplicationReviewPage/ApplicationReviewPage.tsx
@@ -32,8 +32,8 @@ import { ExperienceForDate } from "~/types/experience";
 import { categorizeSkill } from "~/utils/skillUtils";
 import ExperienceTreeItems from "~/components/ExperienceTreeItems/ExperienceTreeItems";
 import ExperienceCard from "~/components/ExperienceCard/ExperienceCard";
+import SkillTree from "~/components/SkillTree/SkillTree";
 
-import SkillTree from "../ApplicationSkillsPage/components/SkillTree";
 import { ApplicationPageProps } from "../ApplicationApi";
 import { useApplicationContext } from "../ApplicationContext";
 import ReviewSection from "./ReviewSection";
