+++ apps/web/src/pages/Applications/ApplicationSkillsPage/ApplicationSkillsPage.tsx
@@ -29,10 +29,10 @@ import {
 } from "~/api/generated";
 import { AnyExperience } from "~/types/experience";
 import { isIncomplete } from "~/validators/profile/skillRequirements";
+import SkillTree from "~/components/SkillTree/SkillTree";
 
 import { ApplicationPageProps } from "../ApplicationApi";
 import { useApplicationContext } from "../ApplicationContext";
-import SkillTree from "./components/SkillTree";
 import SkillDescriptionAccordion from "./components/SkillDescriptionAccordion";
 import useApplicationId from "../useApplicationId";
