+++ apps/web/src/pages/PoolCandidates/ViewPoolCandidatePage/components/ApplicationStatusForm/AssetSkillsFiltered.tsx
@@ -4,7 +4,7 @@ import { useIntl } from "react-intl";
 import { Experience, Skill } from "@gc-digital-talent/graphql";
 import Well from "@gc-digital-talent/ui/src/components/Well";
 
-import SkillTree from "~/pages/Applications/ApplicationSkillsPage/components/SkillTree";
+import SkillTree from "~/components/SkillTree/SkillTree";
 import { getExperiencesSkillIds } from "~/utils/skillUtils";
 
 interface AssetSkillsFilteredProps {
